requirements.txt
aiogram==3.4.0
sqlalchemy==2.0.28
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.1
aiohttp==3.9.3